requirements.txt
astroid==2.4.2
autopep8==1.5.4
Click==7.0
cycler==0.10.0
decorator==4.4.2
flake8==3.8.4
Flask==1.0.2
imageio==2.9.0
isort==5.6.4
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.10
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mccabe==0.6.1
mingus==0.6.0
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
parso==0.7.1
Pillow==8.0.1
pluggy==0.13.1
pycodestyle==2.6.0
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.1
PyWavelets==1.1.1
rope==0.18.0
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
snowballstemmer==2.0.0
tifffile==2020.10.1
toml==0.10.2
ujson==4.0.1
uWSGI==2.0.18
Werkzeug==0.15.5
wrapt==1.12.1
yapf==0.30.0
